+++ src/components/Providers/Providers.tsx
@@ -1,11 +1,16 @@
 "use client";
 
+import { CartProvider } from "@/lib/CartContext";
 import { SessionProvider } from "next-auth/react";
 interface Props {
   children?: React.ReactNode;
 }
 function Providers({ children }: Props) {
-  return <SessionProvider>{children}</SessionProvider>;
+  return (
+    <SessionProvider>
+      <CartProvider>{children}</CartProvider>
+    </SessionProvider>
+  );
 }
 
 export default Providers;
